# swagger.yaml
openapi: 3.0.0
info:
  title: Attraction API
  description: Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.
  version: 0.1.9
servers:
  - url: http://localhost:5000
    description: Optional server description, e.g. Main (production) server
  - url: http://staging-api.example.com
    description: Optional server description, e.g. Internal staging server for testing
paths:
  /attractions:
    get:
      summary: Returns a list of Attraction.
      description: Optional extended description in CommonMark or HTML.
      responses:
        "200": # status code
          description: A JSON array of user names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
    post:
      summary: Creates a attractions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                detail:
                  type: string
                coverimage:
                  type: string
                latitude:
                  type: number
                  format: float
                longitude:
                  type: number
                  format: float
      responses:
        "201":
          description: Created
    put:
      summary: Updated a attractions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: number
                name:
                  type: string
                detail:
                  type: string
                coverimage:
                  type: string
                latitude:
                  type: number
                  format: float
                longitude:
                  type: number
                  format: float
      responses:
        "201":
          description: Created

  /attractions/{id}:
    get:
      summary: Returns a attractions by ID.
      parameters:
        - name: id
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: OK
    delete:
      summary: Delete a attractions by ID.
      parameters:
        - name: id
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: OK